# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BriefAllergy": {
        "properties": {
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "substance": {
            "title": "Substance",
            "type": "string"
          }
        },
        "required": [
          "substance"
        ],
        "title": "BriefAllergy",
        "type": "object"
      },
      "BriefCondition": {
        "properties": {
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "condition"
        ],
        "title": "BriefCondition",
        "type": "object"
      },
      "BriefFlag": {
        "properties": {
          "severity": {
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "enum": [
              "abnormal_lab",
              "needs_review",
              "missing_data",
              "conflict"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "text",
          "type"
        ],
        "title": "BriefFlag",
        "type": "object"
      },
      "BriefLab": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "flag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag"
          },
          "previous": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          },
          "test": {
            "title": "Test",
            "type": "string"
          },
          "trend": {
            "anyOf": [
              {
                "enum": [
                  "up",
                  "down",
                  "stable"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "test",
          "value"
        ],
        "title": "BriefLab",
        "type": "object"
      },
      "BriefMedication": {
        "properties": {
          "dose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose"
          },
          "frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "name"
        ],
        "title": "BriefMedication",
        "type": "object"
      },
      "BriefPriority": {
        "properties": {
          "level": {
            "enum": [
              "routine",
              "review_soon"
            ],
            "title": "Level",
            "type": "string"
          },
          "reasons": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "level"
        ],
        "title": "BriefPriority",
        "type": "object"
      },
...
      "DoctorBriefDTO": {
        "properties": {
          "active_conditions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BriefCondition"
            },
            "title": "Active Conditions",
            "type": "array"
          },
          "active_medications": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BriefMedication"
            },
            "title": "Active Medications",
            "type": "array"
          },
          "allergies": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BriefAllergy"
            },
            "title": "Allergies",
            "type": "array"
          },
          "brief_id": {
            "title": "Brief Id",
            "type": "string"
          },
          "chief_concern": {
            "title": "Chief Concern",
            "type": "string"
          },
          "confidence_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Notes"
          },
          "consult_room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consult Room"
          },
          "flags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BriefFlag"
            },
            "title": "Flags",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "one_line": {
            "title": "One Line",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BriefPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_labs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BriefLab"
            },
            "title": "Recent Labs",
            "type": "array"
          },
          "referral_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Reason"
          },
          "referred_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referred By"
          },
          "source_documents": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Source Documents",
            "type": "array"
          },
          "specialist_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialist Type"
          },
          "suggested_questions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Suggested Questions",
            "type": "array"
          },
          "timeline": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/app__schemas__brief__TimelineEvent"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "brief_id",
          "patient_id",
          "generated_at",
          "model",
          "one_line",
          "chief_concern"
        ],
        "title": "DoctorBriefDTO",
        "type": "object"
      },
...
      "EsanjeewaniExportDTO": {
        "properties": {
          "brief_id": {
            "title": "Brief Id",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "brief_id",
          "text"
        ],
        "title": "EsanjeewaniExportDTO",
        "type": "object"
      },
      "FhirExportDTO": {
        "properties": {
          "brief_id": {
            "title": "Brief Id",
            "type": "string"
          },
          "bundle": {
            "additionalProperties": true,
            "title": "Bundle",
            "type": "object"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "brief_id",
          "bundle"
        ],
        "title": "FhirExportDTO",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PatientCreateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "lang_pref": {
            "default": "mr",
            "title": "Lang Pref",
            "type": "string"
          }
        },
        "title": "PatientCreateRequest",
        "type": "object"
      },
      "PatientDTO": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lang_pref": {
            "default": "mr",
            "title": "Lang Pref",
            "type": "string"
          },
          "patient_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Token"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "PatientDTO",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__schemas__brief__TimelineEvent": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "event": {
            "title": "Event",
            "type": "string"
          }
        },
        "required": [
          "date",
          "event"
        ],
        "title": "TimelineEvent",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Supabase access token (required when SUPABASE_ENABLED=true)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SETU API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/patients": {
      "post": {
        "description": "Legacy anonymous patient create — disabled when Supabase auth is enabled.",
        "operationId": "create_patient_api_v1_patients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Patient",
        "tags": [
          "patients"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/brief": {
      "get": {
        "operationId": "get_brief_api_v1_patients__patient_id__brief_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorBriefDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Brief",
        "tags": [
          "brief"
        ]
      },
      "post": {
        "operationId": "regenerate_brief_api_v1_patients__patient_id__brief_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorBriefDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Regenerate Brief",
        "tags": [
          "brief"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/brief/exports/esanjeewani": {
      "get": {
        "operationId": "get_brief_esanjeewani_api_v1_patients__patient_id__brief_exports_esanjeewani_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EsanjeewaniExportDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Brief Esanjeewani",
        "tags": [
          "brief"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/brief/exports/esanjeewani/download": {
      "get": {
        "operationId": "download_brief_esanjeewani_api_v1_patients__patient_id__brief_exports_esanjeewani_download_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Brief Esanjeewani",
        "tags": [
          "brief"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/brief/fhir": {
      "get": {
        "operationId": "get_brief_fhir_api_v1_patients__patient_id__brief_fhir_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirExportDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Brief Fhir",
        "tags": [
          "brief"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/brief/fhir/download": {
      "get": {
        "operationId": "download_brief_fhir_api_v1_patients__patient_id__brief_fhir_download_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Brief Fhir",
        "tags": [
          "brief"
        ]
      }
    },
...
  }
}